Refactor app to make it more structural
